backend/requirements.txt
Django==5.1.7
djangorestframework # DRF
django-cors-headers # Cross-Origin-Resource-support
markdown            # Markdown support for the browsable API.
django-filter       # Filtering support
psycopg2            # Postgress support
pillow              # Image support
